Rename package "model.dto" to "model"
