+++ src/main/java/org/example/controller/AddTimerController.java
@@ -7,7 +7,7 @@ import javafx.scene.control.SpinnerValueFactory;
 import javafx.scene.control.TextField;
 import javafx.stage.Stage;
 import org.example.service.TimerServiceFactory;
-import org.example.model.dto.TimerCreate;
+import org.example.model.TimerCreate;
 import org.example.service.interfaces.TimerService;
 
 import java.util.concurrent.ExecutionException;
