+++ src/main/java/org/example/controller/TimerController.java
@@ -3,7 +3,7 @@ package org.example.controller;
 import javafx.fxml.FXML;
 import javafx.scene.control.*;
 import org.example.service.TimerServiceFactory;
-import org.example.model.dto.TimerDetails;
+import org.example.model.TimerDetails;
 import org.example.service.interfaces.TimerService;
 import org.example.manager.TimerGuiManager;
 import org.example.manager.TimerListManager;
