+++ src/main/java/org/example/model/AuthRequest.java
@@ -1,4 +1,4 @@
-package org.example.model.dto;
+package org.example.model;
 
 public class AuthRequest {
     private String email;
