+++ src/main/java/org/example/service/AuthServiceImpl.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import org.example.infrastructure.ApiClient;
-import org.example.model.dto.AuthRequest;
+import org.example.model.AuthRequest;
 import org.example.service.interfaces.AuthService;
 
 import java.util.concurrent.CompletableFuture;
