+++ src/main/java/org/example/service/RemoteTimerService.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import org.example.infrastructure.ApiClient;
-import org.example.model.dto.TimerCreate;
-import org.example.model.dto.TimerDetails;
-import org.example.model.dto.TimerUpdate;
+import org.example.model.TimerCreate;
+import org.example.model.TimerDetails;
+import org.example.model.TimerUpdate;
 import org.example.service.interfaces.TimerService;
 
 import java.util.concurrent.CompletableFuture;
